python/eg_nameko/requirements.txt
flasgger==0.5.13
flask-mongoengine==0.7.5
nameko==2.3.1
